+++ pyproject.toml
@@ -86,10 +86,15 @@ lint.ignore = [
 "src/rapids_singlecell/decoupler_gpu/_method_mlm.py" = ["PLR0917"]
 "src/rapids_singlecell/decoupler_gpu/_method_wsum.py" = ["PLR0917"]
 
-
 [tool.ruff.lint.isort]
 known-first-party = ["rapids_singlecell"]
 required-imports = ["from __future__ import annotations"]
+
+[tool.pytest.ini_options]
+addopts = [
+    "--import-mode=importlib",
+    "--strict-markers",
+]
 
 [tool.hatch.envs.hatch-test]
 features = ["test", "rapids12"]
